moved show splash
